Support Chinese comment and ime input 1. Add Chinese charactor input for chess records; 2. Add cursor following for the input method window; 3. Fixed a few bugs in x64 mode

+++ src/ChessBoardBitmap.cpp
@@ -1875,9 +1875,9 @@ void ChessBoardBitmap::PutEx( char piece,
             for( j=0; j < width_bytes/(8*density); j++ )
             {
                 if( *mask++ != '0' )
-                    memcpy( dst, src, density );
-                src += density;
-                dst += density;
+                    memcpy( dst, src, (byte)density );
+                src +=(byte) density;
+                dst +=(byte) density;
             }
         }
     }
